renamed files, changed photo imports, teammate cards
